pyproject.toml
[project]
name = "arxiv-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.119.0",
    "marker-pdf>=1.10.1",
    "markitdown[all]>=0.1.3",
    "pymupdf4llm>=0.0.27",
    "torch>=2.9.0",
    "torchvision>=0.24.0",
    "uvicorn>=0.37.0",
]

